+++ src/Interfaces/PaymentServiceInterface.php
@@ -45,20 +45,21 @@ interface PaymentServiceInterface
     /**
      * Performs an Authorization transaction and returns the resulting Authorization resource.
      *
-     * @param float                  $amount        The amount to authorize.
-     * @param string                 $currency      The currency of the amount.
-     * @param string|BasePaymentType $paymentType   The PaymentType object or the id of the PaymentType to use.
-     * @param string                 $returnUrl     The URL used to return to the shop if the process requires leaving it.
-     * @param Customer|string|null   $customer      The Customer object or the id of the customer resource to reference.
-     * @param string|null            $orderId       A custom order id which can be set by the merchant.
-     * @param Metadata|null          $metadata      The Metadata object containing custom information for the payment.
-     * @param Basket|null            $basket        The Basket object corresponding to the payment.
-     *                                              The Basket object will be created automatically if it does not exist
-     *                                              yet (i.e. has no id).
-     * @param bool|null              $card3ds       Enables 3ds channel for credit cards if available. This parameter is
-     *                                              optional and will be ignored if not applicable.
-     * @param string|null            $invoiceId     The external id of the invoice.
-     * @param string|null            $referenceText A reference text for the payment.
+     * @param float                  $amount         The amount to authorize.
+     * @param string                 $currency       The currency of the amount.
+     * @param string|BasePaymentType $paymentType    The PaymentType object or the id of the PaymentType to use.
+     * @param string                 $returnUrl      The URL used to return to the shop if the process requires leaving it.
+     * @param Customer|string|null   $customer       The Customer object or the id of the customer resource to reference.
+     * @param string|null            $orderId        A custom order id which can be set by the merchant.
+     * @param Metadata|null          $metadata       The Metadata object containing custom information for the payment.
+     * @param Basket|null            $basket         The Basket object corresponding to the payment.
+     *                                               The Basket object will be created automatically if it does not exist
+     *                                               yet (i.e. has no id).
+     * @param bool|null              $card3ds        Enables 3ds channel for credit cards if available. This parameter is
+     *                                               optional and will be ignored if not applicable.
+     * @param string|null            $invoiceId      The external id of the invoice.
+     * @param string|null            $referenceText  A reference text for the payment.
+     * @param string|null            $recurrenceType Recurrence type used for recurring payment.
      *
      * @return Authorization The resulting object of the Authorization resource.
      *
@@ -76,7 +77,8 @@ interface PaymentServiceInterface
         $basket = null,
         $card3ds = null,
         $invoiceId = null,
-        $referenceText = null
+        $referenceText = null,
+        $recurrenceType = null
     ): Authorization;
 
     /**
@@ -96,6 +98,8 @@ interface PaymentServiceInterface
      *                                                 optional and will be ignored if not applicable.
      * @param string|null            $invoiceId        The external id of the invoice.
      * @param string|null            $paymentReference A reference text for the payment.
+     * @param string|null            $recurrenceType   Recurrence type used for recurring payment.
+     *                                                 See \UnzerSDK\Constants\RecurrenceTypes to find all supported types.
      *
      * @return Charge The resulting object of the Charge resource.
      *
@@ -113,7 +117,8 @@ interface PaymentServiceInterface
         $basket = null,
         $card3ds = null,
         $invoiceId = null,
-        $paymentReference = null
+        $paymentReference = null,
+        $recurrenceType = null
     ): Charge;
 
     /**