+++ test/unit/Resources/TransactionTypes/AuthorizationTest.php
@@ -35,6 +35,7 @@ use UnzerSDK\Resources\TransactionTypes\Cancellation;
 use UnzerSDK\Resources\TransactionTypes\Charge;
 use UnzerSDK\test\BasePaymentTest;
 use RuntimeException;
+use function Webmozart\Assert\Tests\StaticAnalysis\object;
 
 class AuthorizationTest extends BasePaymentTest
 {