+++ CHANGELOG.md
@@ -2,8 +2,7 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
-## [1.1.3.1](https://github.com/unzerdev/php-sdk/compare/1.1.3.0..1.1.3.1)
+## [1.1.4.1](https://github.com/unzerdev/php-sdk/compare/1.1.4.0..1.1.4.1)
 ### Added
 *   Added Apple Pay example.
 
@@ -12,6 +11,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 *   Several minor improvements.
+
+## [1.1.4.0](https://github.com/unzerdev/php-sdk/compare/1.1.3.0..1.1.4.0)
+### Added
+*   Enable recurrence type to be set for `charge`, `authorize` and `activateRecurringPayment` methods.
+
+### Changed
+* Enable recurring examples (card paypal)to trigger subsequent transaction from success page.
+* Enable card recurring example to use recurrence type.
+* Several minor improvements.
 
 ## [1.1.3.0](https://github.com/unzerdev/php-sdk/compare/1.1.2.0..1.1.3.0)
 ### Added